requirements.txt
flask
supabase
python-dotenv
pyjwt
bcrypt